added Gitflow, added custom shortcuts, fixed gnus configuration
Added the Magit Gitflow package to Emacs.

Also added some custom shortcuts, such as `SPC o x t f` which is an alias to
`C-u M-q` or `M-1 M-q`. This is due to the change of editing style in Spacemacs,
I went from the Emacs editing style to the hybrid editing style —that is, in
normal mode, I am using evil-mode, and in insert mode, I am using Emacs’
keybinding.

This also lead me to add a couple of shortcuts for Gnus, which had a broken IMAP
configuration (wrong port).