Shortcuts update, better way to exit i3
With ly as my login service, `i3-msg exit` didn’t work anymore. So, I made it so
the service of ly gets restarted, which closes i3.

I also replaced all shortcuts referencing Chromium with their equivalent with
Firefox. I made the switch due to a rencent Chromium update which made my
computer very laggy, while Firefox doesn’t seem to have this problem so far.

I also added a shortcut for toggling the sticky mode for windows, which should
prove useful if I wish to keep certain windows displayed regardless of the
workspace I am on.

Lastly, I added a shortcut to the new feature of `rofi-pass` I implemented,
which allows me to select a password that will be typed for me.