Switched to LSP backend, added packages to install and configuration
In Emacs, I switched to the LSP backend for C/C++, CSS, Dart, HTML, Go,
Javascript and Python. I also removed a redundant line for web-beautify, and I
removed the plantuml layer since I don’t use it.

This lead to the installation of multiple packages, some of which were already
required but not explicitly installed by the bootstrap. I added the packages
existing in the Arch repos and the AUR, and some packages will get installed by
pip or by go’s package manager. I also set up the configuration of Chicken, and
more precisely, of its documentation.