Updated some paths, fixed some code style, added back some shortcuts
Updated some paths after the Emacs configuration directory was moved

Some variables in my dotspacemacs were renamed in order to follow the
Spacemacs coding style with a ~//~ used for marking private variables,
and I prefixed their name with ~dotspacemacs-~ in order to avoid any
collision with other packages or configuration I write.

Added back M-« and M-» in order to fix the M-< and M-> shortcuts issue
with the bépo layout.