[Bin] Remove shebang from source, add it as header argument
This commit cleans up the source blocks of bin.org by removing the
shebangs from the actual source blocks. They are now passed as header
arguments, making the exported files automatically executable. This
removes the need for the two functions that were in the Introduction
header making tangled files executable or not.