[Emacs] Change font to DejaVu Sans Mono, better mu4e visual config
This commit changes the font of Emacs from Fantasque Sans Mono to
DejaVu Sans Mono. The latter has a better support for the
International Phonetic Alphabet (IPA) than the former. Font size is
also adjusted.

In the mu4e header view, the width of the header view when a message
is opened is now better handled since it will be calculated
accordingly to the frame and not the window. In Emacs, the frame is
actually what we commonly call a window, and a window is the area in
which buffers are displayed (damn this pre-GUI terminology in Emacs).
There is a bug with Emacs launched as a server though: since it is
launched as a headless service at first, the edited code cannot know
what the size of the frame will be, so the user will have to
re-evaluate this code snippet once the first frame is created, or when
the frame is resized and the user wishes to see this value updated.

A possible fix for this could be to hook this evaluation to when a
message is opened instead of when Emacs spins up.