[Emacs] Fix evaluation issue of `mu4e-headers-visible-columns' eval
This commit fixes a bug introduced in 8c6e622a7c and 69dc8c7d7c where
the mu4e-headers view’s width was set to a dynamic, frame-dependant
value that could not be computed with Emacs launched as a headless
server.

This commit makes this value evaluate when the `mu4e-headers' mode is
activated, and hence it should get the correct width of the current
frame.