[Emacs] Mu4e react to window width rather than frame width
Referring to the previous commit, it is actually better if the mu4e
headers view sets its own width _when a message is opened_ and not
when the headers view itself is opened. This commit also makes it so
that the width is evaluated based on the width of the window of the
headers view’s window rather than the width of the frame since the
message will be opened in a new window that will spawn in the headers
view's window.