[Emacs] Code style, prompt fix for Eshell
A bug was introduced in b97bbf8 with the fix for Eshell, where the
abbreviated path was not used anymore. This commit fixes it.

This commit also introduces the new macro `phundrak/var-or-if-nil'
which is documented in the code. It has replaced some code already.

The coding style of some recent functions have been updated too,
arguments names now begin with a dollar `$' sign. This is just stylistic
and it has no influence on the code whatsoever.

Finally, some nested `if's in `phundrak/abbr-pwd' were converted in a
single `cond', which led to renaming two variables in order to avoid a
collision between the variable and the function `push' – `pull' has been
renamed accordingly.