[Emacs] Rename `eshell/my-prompt', simplify `phundrak/abbr-pwd'
This commit renames `eshell/my-prompt' to `phundrak/eshell-prompt'.

`phundrak/abbr-pwd' is now simplified, and it is no longer responsible
for fetching the current path which must now be passed as its first
argument. It will also reject any argument which is neither a string
(should be the only type to be passed for public use) nor a list (for
internal usage).