[Emacs] Better code style, replace macro with function
`if's that had one branch are replaced with `when', and `if's which only
actual branch is the else branch are replaced by `unless'.

`phundrak/eshell-git-status' no longer verifies whether its `$path'
argument points to a git repository, that is now the role of the callee.

`phundrak/eshell-prompt' now makes use of more variables to make the
code more readable.

Some documentation is now formatted correctly