[Emacs] Update eshell aliases
Update eshell aliases

I should replace this file later with some Elisp code in the user init
to read aliases from Fish abbreviations.

Something good would be to try to write some Elisp code to emulate Fish
abbreviations, but I have to find what hook or function is called when a
new character is entered in the shell that triggers autocompletion. Code
in this package might help:
https://github.com/szermatt/emacs-bash-completion