[Emacs] Add Swift layer with ob-swift, keybinding to woman
This commit adds to my configuration the Swift layer from Spacemacs, as
well as ob-swift for evaluating Swift code in org-mode –because
apparently it is not part of the layer itself.

The keybinding to `wttrin' is also modified to `SPC o a W', with its
original keybinding `SPC o a w' set to `helm-man-woman', a helm
interface to Emacs’ `woman'.