[Emacs] Remove w3m layer, compile configuration files when exported
The w3m layer is now removed as I did not use it at all.

The conlanging layer is moved to ~/fromGIT/emacs-packages and now a
symlink points to its new location. This is done in order to prevent
Emacs from compiling its .el files when recompiling its configuration
files.

Configuration files are now compiled when emacs.org is tangled by Emacs
itself or if the .el files are newer than the .elc files.

TODO: find a way to natively compile and load these files instead of
just byte-compile them if native compilation is available