[Emacs] Config now correctly tangles without errors
Previously, some code relied on dash.el to generate some code when
tangling my Emacs config, but if Emacs was to be launched without an
up to date tangled configuration, it would attempt to automatically
tangle it. However, due to dash.el not being loaded, it could not find
some of the required functions.

This commit replaces these functions with handmade functions that can
be loaded at tangle-time, and tangling always works.