[Emacs] Fix mu4e keybinds, handle screen width better, and stop spam
For some reason, some keybinds in mu4e weren’t getting undefined
anymore, this is now fixed.

Add automatic vertical split if the screen is considered wide
enough (120 characters for the headers mode and 80 characters for the
message view mode). If the screen or the headers mode window are not
wide enough, the window will split horizontally instead of vertically.

This commit also adds a function to quickly insert a response to
spammers to get them trapped in an infinite loop of emails. See the
new links in this commit.