[StumpWM] New keybinds, heading renamed
Keybinds related to the number row can now be automatically generated.
Related to this, keybinds to move windows between groups, following
them or not, and merge groups into the current one are added with this
commit.

The application keymap is now moved to the root map in order to keep
some consistency with my Emacs keybinds.