[StumpWM] Add group, change group order, SYS now dynamic group
Add a new group dedicated to gaming, making it the 8th group.

DEV and SYS were swapped, and now SYS is a dynimacally tiling group.

Groups can now be defined as a `Dynamic' or `Floating' group. Any
other value in the corresponding column will be considered to be a
manually tiling group.