docs(hyprland, scripts): better screenshot script
This commit adds the possibility of using swappy to edit screenshots.
It also changes the behaviour of the delay by first allowing the user
to select the area they want to capture, and then only does it wait
for the indicated amount of time before it takes the screen capture.